pyproject.toml
[project]
name = "brasileirao-analise"
version = "0.1.0"
description = "Projeto para analisar os resultados do campeonato brasileiro"
authors = [
    {name = "Joao Pedro Santos",email = "jpedro.csantos@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "fastparquet (>=2024.11.0,<2025.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "mplcursors (>=0.6,<0.7)",
    "wordcloud (>=1.9.4,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "dash (>=2.18.2,<3.0.0)",
    "plotly (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
